LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.textRendering.settings-expected.txt

2d.text.drawing.style.textRendering.settings
Testing basic functionalities of textRendering in Canvas
Actual output:

FAIL Testing basic functionalities of textRendering in Canvas assert_equals: ctx.textRendering === "Auto" (got [undefined], expected Auto[string]) expected (string) "Auto" but got (undefined) undefined
